\name{QuantileDSPM-package}
\alias{QuantileDSPM-package}
\alias{QuantileDSPM}
\docType{package}
\title{\packageTitle{QuantileDSPM}}
\description{\packageDescription{QuantileDSPM}}
\section{Package Content}{\packageIndices{QuantileDSPM}}
\author{\packageAuthor{QuantileDSPM}}
\section{Maintainer}{\packageMaintainer{QuantileDSPM}}
\keyword{package}
